+++ ydb/core/tx/columnshard/data_reader/contexts.cpp
@@ -1,0 +1,4 @@
+#include "contexts.h"
+
+namespace NKikimr::NOlap::NDataFetcher {
+}   // namespace NKikimr::NOlap::NDataFetcher
