+++ ydb/core/tx/columnshard/data_reader/contexts.h
@@ -1,0 +1,204 @@
+#pragma once
+#include <ydb/core/tx/columnshard/blobs_reader/task.h>
+#include <ydb/core/tx/columnshard/data_accessor/manager.h>
+#include <ydb/core/tx/columnshard/engines/portions/data_accessor.h>
+#include <ydb/core/tx/limiter/grouped_memory/usage/abstract.h>
+#include <ydb/core/tx/limiter/grouped_memory/usage/service.h>
+
+#include <ydb/library/accessor/accessor.h>
+
+namespace NKikimr::NOlap::NDataFetcher {
+
+enum class EFetchingStage : ui32 {
+    Created = 0,
+    AskAccessorResources,
+    AskDataResources,
+    AskAccessors,
+    ReadBlobs,
+    Finished,
+    Error
+};
+
+class TCurrentContext {
+private:
+    std::optional<std::vector<TPortionDataAccessor>> Accessors;
+    std::vector<std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>> ResourceGuards;
+    std::shared_ptr<NGroupedMemoryManager::TProcessGuard> MemoryProcessGuard;
+    std::shared_ptr<NGroupedMemoryManager::TScopeGuard> MemoryProcessScopeGuard;
+    std::shared_ptr<NGroupedMemoryManager::TGroupGuard> MemoryProcessGroupGuard;
+    static inline TAtomicCounter Counter = 0;
+    const ui64 MemoryProcessId = Counter.Inc();
+    std::optional<NBlobOperations::NRead::TCompositeReadBlobs> Blobs;
+
+public:
+    ui64 GetMemoryProcessId() const {
+        return MemoryProcessId;
+    }
+
+    void SetBlobs(NBlobOperations::NRead::TCompositeReadBlobs&& blobs) {
+        AFL_VERIFY(!Blobs);
+        Blobs = std::move(blobs);
+    }
+
+    NBlobOperations::NRead::TCompositeReadBlobs& MutableBlobs() {
+        AFL_VERIFY(!!Blobs);
+        return *Blobs;
+    }
+
+    void ResetBlobs() {
+        Blobs.reset();
+    }
+
+    TCurrentContext() {
+        MemoryProcessGuard = NGroupedMemoryManager::TCompMemoryLimiterOperator::BuildProcessGuard(
+            MemoryProcessId, { std::make_shared<NGroupedMemoryManager::TStageFeatures>("DEFAULT", 1000000000, 5000000000, nullptr, nullptr) });
+        MemoryProcessScopeGuard = NGroupedMemoryManager::TCompMemoryLimiterOperator::BuildScopeGuard(MemoryProcessId, 1);
+        MemoryProcessGroupGuard = NGroupedMemoryManager::TCompMemoryLimiterOperator::BuildGroupGuard(MemoryProcessId, 1);
+    }
+
+    void SetPortionAccessors(std::vector<TPortionDataAccessor>&& acc) {
+        AFL_VERIFY(!Accessors);
+        Accessors = std::move(acc);
+    }
+
+    const std::vector<TPortionDataAccessor> GetPortionAccessors() const {
+        AFL_VERIFY(Accessors);
+        return *Accessors;
+    }
+
+    void RegisterResourcesGuard(std::shared_ptr<NGroupedMemoryManager::TAllocationGuard>&& g) {
+        ResourceGuards.emplace_back(std::move(g));
+    }
+};
+
+class IFetchCallback {
+private:
+    virtual void DoOnFinished(TCurrentContext&& context) = 0;
+    virtual void DoOnError(const TString& errorMessage) = 0;
+    bool IsFinished = false;
+
+public:
+    void OnFinished(TCurrentContext&& context) {
+        AFL_VERIFY(!IsFinished);
+        IsFinished = true;
+        return DoOnFinished(std::move(context));
+    }
+
+    void OnError(const TString& errorMessage) {
+        AFL_VERIFY(!IsFinished);
+        IsFinished = true;
+        return DoOnError(errorMessage);
+    }
+};
+
+class TEnvironment {
+private:
+    std::shared_ptr<NDataAccessorControl::IDataAccessorsManager> DataAccessorsManager;
+    std::shared_ptr<IStoragesManager> StoragesManager;
+
+public:
+};
+
+class TPortionsDataFetcher;
+
+class IFetchingStep {
+public:
+    enum class EStepResult {
+        Continue,
+        Detached,
+        Error
+    };
+
+private:
+    virtual EStepResult DoExecute(const std::shared_ptr<TPortionsDataFetcher>& fetchingContext) const = 0;
+
+public:
+    [[nodiscard]] EStepResult Execute(const std::shared_ptr<TPortionsDataFetcher>& fetchingContext) const {
+        return DoExecute(fetchingContext);
+    }
+};
+
+class TScript {
+private:
+    const TString ClassName;
+    std::vector<std::shared_ptr<IFetchingStep>> Steps;
+
+public:
+    const TString& GetClassName() const {
+        return ClassName;
+    }
+
+    TScript(std::vector<std::shared_ptr<IFetchingStep>>&& steps, const TString& className)
+        : ClassName(className)
+        , Steps(std::move(steps)) {
+    }
+
+    const std::shared_ptr<IFetchingStep>& GetStep(const ui32 index) const {
+        AFL_VERIFY(index < Steps.size());
+        return Steps[index];
+    }
+
+    ui32 GetStepsCount() const {
+        return Steps.size();
+    }
+};
+
+class TScriptExecution {
+private:
+    std::shared_ptr<TScript> Script;
+    ui32 StepIndex = 0;
+
+public:
+    TScriptExecution(const std::shared_ptr<TScript>& script)
+        : Script(script) {
+        AFL_VERIFY(Script);
+    }
+
+    const TString& GetScriptClassName() const {
+        return Script->GetClassName();
+    }
+
+    const std::shared_ptr<IFetchingStep>& GetCurrentStep() const {
+        return Script->GetStep(StepIndex);
+    }
+
+    bool IsFinished() const {
+        return StepIndex == Script->GetStepsCount();
+    }
+
+    void Next() {
+        AFL_VERIFY(!IsFinished());
+        ++StepIndex;
+    }
+};
+
+class TFullPortionInfo {
+private:
+    YDB_READONLY_DEF(std::shared_ptr<TPortionInfo>, PortionInfo);
+    YDB_READONLY_DEF(ISnapshotSchema::TPtr, Schema);
+
+public:
+    TFullPortionInfo(const std::shared_ptr<TPortionInfo>& portionInfo, const ISnapshotSchema::TPtr& schema)
+        : PortionInfo(portionInfo)
+        , Schema(schema) {
+    }
+};
+
+class TRequestInput {
+private:
+    YDB_READONLY_DEF(std::vector<std::shared_ptr<TFullPortionInfo>>, Portions);
+    YDB_READONLY(NBlobOperations::EConsumer, Consumer, NBlobOperations::EConsumer::UNDEFINED);
+    YDB_READONLY_DEF(TString, ExternalTaskId);
+
+public:
+    TRequestInput(const std::vector<std::shared_ptr<TPortionInfo>>& portions, const std::shared_ptr<TVersionedIndex>& versions,
+        const NBlobOperations::EConsumer consumer, const TString& externalTaskId)
+        : Consumer(consumer)
+        , ExternalTaskId(externalTaskId) {
+        for (auto&& i : portions) {
+            Portions.emplace_back(TFullPortionInfo(i, versions->GetSchemaVerified(i->GetSchemaVersionVerified())));
+        }
+    }
+};
+
+}   // namespace NKikimr::NOlap::NDataFetcher
