+++ ydb/core/tx/columnshard/data_reader/fetching_steps.cpp
@@ -1,0 +1,4 @@
+#include "fetching_steps.h"
+
+namespace NKikimr::NOlap::NDataFetcher {
+}   // namespace NKikimr::NOlap::NDataFetcher
