+++ ydb/core/tx/columnshard/data_reader/ya.make
@@ -3,6 +3,9 @@ LIBRARY()
 SRCS(
     actor.cpp
     fetcher.cpp
+    fetching_executor.cpp
+    fetching_steps.cpp
+    contexts.cpp
 )
 
 PEERDIR(
